chore: :truck: rename sub-package to match eventual intention
