+++ src/seedcase_sprout/core/check_datapackage/check_package_properties.py
@@ -1,11 +1,11 @@
-from seedcase_sprout.core.checks.add_package_recommendations import (
+from seedcase_sprout.core.check_datapackage.add_package_recommendations import (
     add_package_recommendations,
 )
-from seedcase_sprout.core.checks.check_error import CheckError
-from seedcase_sprout.core.checks.check_object_against_json_schema import (
+from seedcase_sprout.core.check_datapackage.check_error import CheckError
+from seedcase_sprout.core.check_datapackage.check_object_against_json_schema import (
     check_object_against_json_schema,
 )
-from seedcase_sprout.core.checks.config import (
+from seedcase_sprout.core.check_datapackage.config import (
     DATA_PACKAGE_SCHEMA_PATH,
 )
 from seedcase_sprout.core.read_json import read_json
