+++ src/seedcase_sprout/core/check_datapackage/schemas/data-package-schema.json
